Add file env.go for checking against env. vars.
This file contains only one function. If it finds a discrepancy between
the internal config state and the value of the environment variables, it
will choose the value of the environment variable. This function also
decides how Arken's environment variables are named:
ARKEN_<CATEGORY>_<DESCRIPTOR>
where category is the name of the substruct of struct config.go::config in
which the field was found, and the descriptor is the name of the field.
 Changes to be committed:
	modified:   config/config.go
	new file:   config/env.go