Add helper functiuons to clean up code
Also make it so that if the given value is not valid and must be changed, this change is
reflected in the global config struct. The new value will be inserted in bytes so as to
avoid floating point values in GB. For example, if the user wants to allocate "max",
Global.General.PoolSize will be changed to the user's hard drive capacity in bytes,
in addition to that value being put in GlobalDiskInfo.