Make DiskInfoProvider access struct by gettters/setters
Remove some unnecesary functions, add a generic utility function for turning bytes into a human-readable string with appropriate units and accompanying tests. Make it so that if the config's pool size had to change, the new value is put in there in GB or TB, through the use of the new DiskInfoProvider method GetPrettyPoolSize(). Usage of this method can be seen in its test function.